Repackage classes into relevant packages
