+++ src/main/java/Dan/task/Event.java
@@ -1,24 +1,24 @@
-package dan;
+package dan.task;
 
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 import java.time.LocalDateTime;
 
 /**
- * One of the task types. A task associated with a timing for which it is due by.
+ * One of the task types. A task associated with a timing for which it's happening
  */
-public class Deadline extends Task {
-    private static final String ICON = "D";
+public class Event extends Task {
+    private static final String ICON = "E";
     protected LocalDateTime dateString;
 
     /**
-     * A task with an associated description and deadline
+     * Creates an Event with its associated description and date.
      *
      * @param description The description of the task
      * @param dateString A String representing a date-time in the pattern "dd/MM/yyyy HHmm"
      * @throws DateTimeParseException If the date-time string is not part of its regular pattern
      */
-    Deadline(String description, String dateString) throws DateTimeParseException {
+    public Event(String description, String dateString) throws DateTimeParseException {
         super(description);
         this.dateString = LocalDateTime.parse(dateString, DateTimeFormatter.ofPattern("dd/MM/yyyy HHmm"));
     }
