+++ src/main/java/Dan/task/TaskList.java
@@ -1,4 +1,4 @@
-package dan;
+package dan.task;
 
 import java.util.List;
 import java.util.ArrayList;
