+++ src/main/java/Dan/tasklistreader/TaskListReader.java
@@ -1,4 +1,4 @@
-package dan;
+package dan.tasklistreader;
 
 import java.nio.file.Files;
 import java.nio.file.Path;
